move proxy settings to sdk
