--- src/notte/browser/__init__.py
@@ -1,8 +1,0 @@
-from typing_extensions import TypedDict
-
-
-class ProxySettings(TypedDict, total=False):
-    server: str
-    bypass: str | None
-    username: str | None
-    password: str | None
